tk suggestions
