+++ tests/__init__.py
@@ -1,4 +1,4 @@
 """EthicML"""
 
 from . import loading_data_test, metrics
-from .metrics import monolith_metric_test
+from .metrics import nonparamaterized_metric_test
